+++ docs/src/release_notes.rst
@@ -12,6 +12,7 @@ Release Notes
 + Updated ``tcprofile`` command to support permutations logic for Apps with layout.json conditional input parameters.
 + Updated ``tcprofile`` command to update the profile schema to v3. **app.arg** is now **app.arg.optional** and **app.arg.required**.
 + Updated ``tcrun`` arg parsing logic to use a **=** separator instead of a space.
++ Updated batch module to support new 5.8+ merge of file hash feature.
 
 0.9.6
 -----